feat(frontend): Introduce reusable layout component to avoid code duplication
