+++ frontend/src/views/OpenApiUI.vue
@@ -1,141 +1,35 @@
-/*-
- * ========================LICENSE_START=================================
- * PREvant Frontend
- * %%
- * Copyright (C) 2018 - 2019 aixigo AG
- * %%
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- * =========================LICENSE_END==================================
- */
 <template>
-   <transition name="modal">
-      <div class="ra-modal-mask">
-         <div class="ra-modal-wrapper">
-            <div class="ra-modal-container">
-
-               <div class="ra-modal-header">
-                  <h1>API Documentation<span v-if="showAdditionalHeadlineInformation"> – {{ $route.params.title }}</span></h1>
-
-                  <span @click="close()">
-                     <font-awesome-icon icon="window-close" class="ra-modal-close-button" aria-label="Close"/>
-                  </span>
-               </div>
-
-               <div class="ra-modal-body">
-                   <div ref="swagger-ui"></div>
-               </div>
-            </div>
-         </div>
-      </div>
-   </transition>
+  <ModalLikeLayout title="API Documentation" :title-suffix="route.query.title">
+    <div class="open-api-ui" ref="asyncapi"></div>
+  </ModalLikeLayout>
 </template>
 
-<style type="text/css">
-   @import "swagger-ui/dist/swagger-ui.css";
+<style lang="css" src="swagger-ui/dist/swagger-ui.css"></style>
 
-   /* Fixes swagger ui response column width */
-   .ra-modal-container .col.response-col_status {
-       width: 10% !important;
-   }
-   .ra-modal-container .col_header.parameters-col_name {
-       width: 10% !important;
-   }
-   .ra-modal-container .col.parameters-col_name {
-       width: 10% !important;
-   }
-
-   .ra-modal-close-button {
-       position: absolute;
-       right: 20px;
-       top: 20px;
-   }
-
-   .ra-modal-mask {
-      position: fixed;
-      z-index: 9998;
-      top: 0;
-      left: 0;
-      width: 100%;
-      height: 100%;
-      background-color: rgba(0, 0, 0, .5);
-      display: table;
-      transition: opacity .3s ease;
-   }
-
-   .ra-modal-wrapper {
-      display: table-cell;
-      vertical-align: middle;
-   }
-
-   .ra-modal-container {
-      position: fixed;
-      top: 50px;
-      left: 50px;
-      bottom: 50px;
-      right: 50px;
-      padding: 20px 30px;
-      background-color: #fff;
-      border-radius: 2px;
-      box-shadow: 0 2px 8px rgba(0, 0, 0, .33);
-      transition: all .3s ease;
-      font-family: Helvetica, Arial, sans-serif;
-
-      overflow-y: scroll;
-   }
-
-   .ra-modal-header h3 {
-      margin-top: 0;
-      color: #42b983;
-   }
-
-   .ra-modal-body {
-      margin: 20px 0;
-   }
-
-   .ra-modal-enter .ra-modal-container,
-   .ra-modal-leave-active .ra-modal-container {
-      -webkit-transform: scale(0.1);
-      transform: scale(0.1);
-   }
+<style lang="scss" scoped>
+/* Fixes swagger ui response column width */
+.open-api-ui {
+  :deep(.response-col_status) {
+    width: 10% !important;
+  }
+  :deep(.parameters-col_name) {
+    width: 10% !important;
+  }
+}
 </style>
 
 <script setup>
+import { onMounted, useTemplateRef } from "vue";
+import { useRoute } from "vue-router";
 import SwaggerUI from "swagger-ui";
-import { computed, onMounted, useTemplateRef } from "vue";
-import { useRoute, useRouter } from "vue-router";
-import { useEscape } from "../composables/useEscape";
+import ModalLikeLayout from "../layouts/ModalLikeLayout.vue";
 
 const route = useRoute();
-const showAdditionalHeadlineInformation = computed(() => route.params.title != null);
-
-const router = useRouter();
-function close() {
-  router.push(router.options.history.state.back ?? "/");
-}
-useEscape(close);
-
-const asyncapi = useTemplateRef("swagger-ui");
+const asyncapi = useTemplateRef("asyncapi");
 onMounted(() => {
   SwaggerUI({
     url: route.params.url,
-    domNode: asyncapi.value
+    domNode: asyncapi.value,
   });
 });
 </script>
-
